requirements.txt
Flask
PyYAML
psutil
telnetlib3
# Note: telnetlib3 is required for Python 3.13+ (telnetlib was removed from stdlib)
# For older Python versions, the standard telnetlib will work
